+++ test/syntax.jl
@@ -819,7 +819,7 @@ let f = function (x; kw...)
 end
 
 # normalization of Unicode symbols (#19464)
-let ε=1, μ=2, x=3, î=4, ⋅=5
+let ε=1, μ=2, x=3, î=4, ⋅=5, (-)=6
     # issue #5434 (mu vs micro):
     @test Meta.parse("\u00b5") === Meta.parse("\u03bc")
     @test µ == μ == 2
@@ -832,6 +832,9 @@ let ε=1, μ=2, x=3, î=4, ⋅=5
     # middot char · or · vs math dot operator ⋅ (#25098)
     @test Meta.parse("\u00b7") === Meta.parse("\u0387") === Meta.parse("\u22c5")
     @test (·) == (·) == (⋅) == 5
+    # minus − vs hyphen-minus - (#26193)
+    @test Meta.parse("\u2212") === Meta.parse("\u002d")
+    @test (−) == (-) == 6
 end
 
 # issue #8925